refactor: move acc_stats to analysis/helpers
